% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ending.R
\docType{data}
\name{ending}
\alias{ending}
\title{Data of lung cancer state}
\format{
A matrix with 192 rows and 2 variables:
\describe{
 \item{ID}{ID of individuals}
 \item{state}{lung cancer state of individuals}
 }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4115}
}
\usage{
ending
}
\description{
Data of lung cancer state
}
\keyword{cancer}
\keyword{data}
\keyword{state}
