% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\docType{data}
\name{expression}
\alias{expression}
\title{Data of gene expression}
\format{
A matrix with 22215 rows and 193 variables:
\describe{
 \item{ID}{probe name}
 \item{GSM93997}{gene expression of GSM93997}
 }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4115}
}
\usage{
expression
}
\description{
Data of gene expression
}
\keyword{expression}
\keyword{gene}
