% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genename.R
\docType{data}
\name{genename}
\alias{genename}
\title{Data of genesymbol}
\format{
A matrix with 66 rows and 2 variables:
\describe{
 \item{FHIT}{genesymbols in hsa05223 pathway of KEGG}
 }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4115}
}
\usage{
genename
}
\description{
Data of genesymbol
}
\keyword{genesymbol}
