% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.R
\docType{data}
\name{pro}
\alias{pro}
\title{Data of probe and genesymbol}
\format{
A matrix with 22215 rows and 2 variables:
\describe{
 \item{ID}{probe name}
 \item{GeneSymbol}{genename}
 }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4115}
}
\usage{
pro
}
\description{
Data of probe and genesymbol
}
\keyword{genesymbol}
\keyword{probe}
